+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * add tracking state machine (support price changes)
 * allow an array of rules in a custom validator
 * extend input syntax to allow using an array of data
+* add searching and transformation of the HTML during extraction
 
 ## 0.2.0
 * add `--cron` option and improve cron setup